mqttgateway: Skip the pollms sleep if data are processed ("Fast Response Mode")
"Fast Response Mode"
In the case that the last cycle processed any udp or mqtt data, skip the delay for the next cycle.
Needs to be tested, will cause HEAVY load on Miniserver and LoxBerry esp. on restart.
Might not come into release with default enabled, but as config parameter.